Reorganize add/remove code
